Moving mixins to a dedicated file
